KTNB-666 Compiler errors now also reference the line causing the problem.
It also fixes an issue in https://youtrack.jetbrains.com/issue/KTNB-245/Make-exceptions-in-cell-outputs-navigable. 

In particular, the case where exceptions happen in generated code that isn't visible to the user. In that case, we do not point to the specific line, but only the cell number.

The heuristic for doing this is a bit naive. It is just checking if the error line is > visible lines of code. A compiler plugin could inject code in the middle of the user code that would mess with this, but the chance is probably low. The change was tested with our ktor-client integration where it does work correctly.

Merge-request: KDS-MR-43
Merged-by: Christian Melchior <christian.melchior@jetbrains.com>