Add support for K2 Repl Mode
This commit adds support for the K2 Repl. K2 Repl Mode is still Experimental and known bugs exists. See https://youtrack.jetbrains.com/issue/OSIP-299/Implement-K2-REPL-for-Kotlin-Notebooks. It is disabled by default and must be enabled through the ReplCompilerMode kernel argument. In the IntelliJ Notebook Plugin, this toggle is in the Kotlin Notebook Settings and is hidden behind the `kotlin.notebook.replCompilerMode.enabled` registry flag.