Use Sonatype Central for publishing
As OSSRH has been deprecated, this PR migrates this repository to use Sonatype Central for publishing. Most of the details are in the kotlin-libs-publisher plugin (new URLs) and on Team City (new access tokens). So the changes to this repo are minor.

Merge-request: KDS-MR-83
Merged-by: Christian Melchior <christian.melchior@jetbrains.com>
(cherry picked from commit 95919e7d7142bbdf5a1ad7ce9dae61dc81962138)